+++ metricflow/dataflow/builder/node_data_set.py
@@ -12,7 +12,7 @@ from metricflow.dataflow.dataflow_plan import (
     DataflowPlanNode,
 )
 from metricflow.dataset.sql_dataset import SqlDataSet
-from metricflow.plan_conversion.dataflow_to_sql import (
+from metricflow.plan_conversion.to_sql_plan.dataflow_to_sql import (
     DataflowNodeToSqlSubqueryVisitor,
 )
